@@ -9,8 +9,8 @@
   "main": "build/index.js",
   "types": "build/index.d.ts",
   "scripts": {
-    "test": "npm run build && jest --coverage",
-    "test-watch": "npm run build && jest --watchAll --coverage",
+    "test": "jest --coverage",
+    "test-watch": "jest --watchAll --coverage",
     "lint": "./node_modules/eslint/bin/eslint.js --ext js,jsx src",
     "lint-fix": "./node_modules/eslint/bin/eslint.js --ext js,jsx src --fix",
     "format:check": "prettier --config ./.prettierrc --list-different \"src/**/*{.ts,.js,.json,.css,.scss}\"",
@@ -30,22 +30,20 @@
     }
   },
   "devDependencies": {
-    "eslint": "^7.17.0",
-    "eslint-config-airbnb": "^17.1.0",
-    "eslint-plugin-compat": "^3.9.0",
-    "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jsx-a11y": "^6.1.2",
-    "eslint-plugin-promise": "^4.0.1",
-    "eslint-plugin-react": "^7.11.1",
-    "husky": "1.1.2",
-    "jest": "^26.6.3",
-    "prettier": "1.14.3",
-    "pretty-quick": "^1.8.0",
+    "@types/jest": "29.5.12",
+    "eslint": "8.57.0",
+    "eslint-plugin-compat": "6.0.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-react": "7.35.0",
+    "husky": "9.1.5",
+    "jest": "29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "prettier": "3.3.3",
+    "pretty-quick": "4.0.0",
     "react": "^16.7.0-alpha.0",
-    "ts-jest": "^26.4.4",
-    "ts-node": "^9.1.1",
-    "typescript": "^4.1.3",
-    "@types/jest": "^26.0.19"
+    "ts-jest": "29.2.4",
+    "ts-node": "10.9.2",
+    "typescript": "5.5.4"
   },
   "engines": {
     "node": ">=8",
@@ -61,6 +59,5 @@
     "type"
   ],
   "author": "haldarmahesh",
-  "license": "MIT",
-  "dependencies": {}
+  "license": "MIT"
 }
